WIP: fetch contrabands from the server
