+++ backend/routes/contraband.routes.ts
@@ -1,0 +1,6 @@
+import { Router } from "express";
+import { getContrabands } from "../controllers/contraband.controller.js";
+
+export const contrabandRouter = Router();
+
+contrabandRouter.get("/", getContrabands);
